+++ isomorphic/index.js
@@ -2,6 +2,7 @@
 export * from 'ctx-core/be'
 export * from 'ctx-core/rmemo'
 export * from './for/index.js'
+export * from './list/index.js'
 export * from './switch/index.js'
 /** @type {browser__relement_T|server__relement_T} */
 export function relement__use(relement) {